+++ app/state/applets/applets.selectors.js
@@ -20,6 +20,8 @@ export const notificationsSelector = R.path(['applets', 'notifications']);
 
 export const invitesSelector = R.path(['applets', 'invites']);
 
+export const activityAccessSelector = R.path(['applets', 'activityAccess']);
+
 export const currentInviteSelector = R.path(['applets', 'currentInvite']);
 
 // Flatten the applet activities into a single list